# Conflicts:
#	Assets/Scripts/Network/Server/Server.cs

+++ Assets/Scripts/Network/Common/PacketHandler/Impl/S2C/PlayerInputPacketHandler.cs
@@ -9,7 +9,7 @@ public class PlayerInputPacketHandler : IPacketHandler
         {
             PlayerInput playerInput = new PlayerInput()
             {
-                ownerId = reader.ReadInt32(),
+                ownerId = reader.ReadInt64(),
                 actionType = (PlayerInput.ActionType)reader.ReadByte(),
                 panelId = reader.ReadByte(),
                 inputElementId = reader.ReadByte(),
